Split the lifecycle management methods off of `_types.go` files.
Now these files excusively have type/enum definitions.

Fixes: https://github.com/knative/serving/issues/3442
